+++ services/catalog/src/simcore_service_catalog/api/rest/__init__.py
@@ -1,12 +1,10 @@
 from fastapi import FastAPI
-from servicelib.fastapi.http_error import set_app_default_http_error_handlers
 
+from .errors import setup_rest_api_error_handlers
 from .routes import setup_rest_api_routes
 
 
 def initialize_rest_api(app: FastAPI):
 
     setup_rest_api_routes(app)
-
-    # defaults error handling
-    set_app_default_http_error_handlers(app)
+    setup_rest_api_error_handlers(app)
