+++ services/catalog/src/simcore_service_catalog/api/rest/_services.py
@@ -20,9 +20,9 @@ from ..._constants import (
     LIST_SERVICES_CACHING_TTL,
     RESPONSE_MODEL_POLICY,
 )
-from ...db.repositories.groups import GroupsRepository
-from ...db.repositories.services import ServicesRepository
 from ...models.services_db import ServiceAccessRightsAtDB, ServiceMetaDataDBGet
+from ...repository.groups import GroupsRepository
+from ...repository.services import ServicesRepository
 from ...service.director import DirectorApi
 from .._dependencies.database import get_repository
 from .._dependencies.director import get_director_api
