+++ services/catalog/src/simcore_service_catalog/api/rpc/_services.py
@@ -23,9 +23,9 @@ from servicelib.rabbitmq.rpc_interfaces.catalog.errors import (
     CatalogForbiddenError,
     CatalogItemNotFoundError,
 )
-from simcore_service_catalog.db.repositories.groups import GroupsRepository
+from simcore_service_catalog.repository.groups import GroupsRepository
 
-from ...db.repositories.services import ServicesRepository
+from ...repository.services import ServicesRepository
 from ...service import services_api
 from .._dependencies.director import get_director_api
